Created Billing Page
If a user skips the free trial and signs up for a plan, it takes them
to the billing page.

Billing page has access to multiform data, and the page validates that
it MUST have the multiform data to display the page.

Next steps are to input the data from page one in hidden input fields,
so that the user can create an account after Stripe verifies that the
payment was submitted.