Refactored Registration Process
The user account is created, and THEN they are directed to billing.
It’s much better this way.

A new view still needs to be made for plan changes, or trial users that
still want to upgrade. If a user is logged in, the pricing table
buttons should redirect to the change plan page instead of the register
page.