Finished Design for Suggestions Page
Some of the API data is pulling correctly as well, such as the data
needed for the charts (from Quandl). Additional data (such as the
suggestions data from the database) still needs to be pulled properly
and returned to the view. But everything after that point seems to be
working.